[ads] Cleanup priority util
